requirements.txt
scratchattach
warp-python
python-dotenv